requirements.txt
python-telegram-bot==13.15
selenium
pyautogui
chromedriver-autoinstaller
urllib3==1.26.5
